Fix overflow

+++ src/ppu/window.rs
@@ -13,7 +13,7 @@ pub struct Window {
 
 impl Window {
     pub fn on(&self, lcd: &Lcd) -> bool {
-        self.is_visible(lcd) && lcd.ly >= self.y && lcd.ly < self.y + LCD_Y_RES
+        self.is_visible(lcd) && lcd.ly >= self.y && (lcd.ly as u16) < (self.y as u16 + LCD_Y_RES as u16)
     }
     pub fn is_visible(&self, lcd: &Lcd) -> bool {
         lcd.control.win_enable() && self.x <= 166 && self.y < LCD_Y_RES
